requirements.txt
av==11.0.0
cbor2==5.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
coloredlogs==15.0.1
ctranslate2==4.7.0
faster-whisper==0.10.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
hf-xet==1.2.0
huggingface_hub==0.36.1
humanfriendly==10.0
idna==3.11
luma.core==2.5.3
luma.oled==3.14.0
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.23.2
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
pvporcupine==3.0.0
PyAudio==0.2.14
PyYAML==6.0.3
requests==2.32.5
setuptools==80.10.2
smbus2==0.6.0
sympy==1.14.0
tokenizers==0.15.2
tqdm==4.67.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
